+++ solvers/LinearElasticSolver/EquationSystems/LinearElasticSystem.cpp
@@ -48,6 +48,14 @@
 #include <StdRegions/StdNodalTriExp.h>
 #include <StdRegions/StdNodalTriExp.h>
 
+#ifdef NEKTAR_USE_MPI
+#include <MultiRegions/GlobalLinSysXxtStaticCond.h>
+#endif
+
+#ifdef NEKTAR_USE_PETSC
+#include <MultiRegions/GlobalLinSysPETScStaticCond.h>
+#endif
+
 namespace Nektar
 {
 
@@ -154,7 +162,7 @@ void LinearElasticSystem::v_InitObject()
             ::AllocateSharedPtr(m_session,
                                 m_graph,
                                 u->GetLocalToGlobalMap(),
-                                m_boundaryConditions,
+                                m_fields[0]->GetBndConditions(),
                                 m_fields);
     }
 
@@ -166,7 +174,7 @@ void LinearElasticSystem::v_InitObject()
             ::AllocateSharedPtr(m_session,
                                 m_graph,
                                 u->GetLocalToGlobalMap(),
-                                m_boundaryConditions,
+                                m_fields[0]->GetBndConditions(),
                                 m_fields);
     }
 
@@ -411,6 +419,26 @@ void LinearElasticSystem::v_DoSolve()
                 key, m_fields[0], m_schurCompl, m_BinvD, m_C, m_Dinv,
                 m_assemblyMap, MultiRegions::NullPreconditionerSharedPtr);
     }
+#ifdef NEKTAR_USE_PETSC
+    else if (m_assemblyMap->GetGlobalSysSolnType() ==
+             MultiRegions::ePETScStaticCond)
+    {
+        linSys = MemoryManager<
+            MultiRegions::GlobalLinSysPETScStaticCond>::AllocateSharedPtr(
+                key, m_fields[0], m_schurCompl, m_BinvD, m_C, m_Dinv,
+                m_assemblyMap);
+    }
+#endif
+#ifdef NEKTAR_USE_MPI
+    else if (m_assemblyMap->GetGlobalSysSolnType() ==
+             MultiRegions::eXxtStaticCond)
+    {
+        linSys = MemoryManager<
+            MultiRegions::GlobalLinSysXxtStaticCond>::AllocateSharedPtr(
+                key, m_fields[0], m_schurCompl, m_BinvD, m_C, m_Dinv,
+                m_assemblyMap);
+    }
+#endif
 
     linSys->Initialise(m_assemblyMap);
 
@@ -680,15 +708,23 @@ void LinearElasticSystem::v_DoSolve()
 
         for (i = 0; i < bndCondExp.num_elements(); ++i)
         {
-            const Array<OneD,const NekDouble> &bndCoeffs =
-                bndCondExp[i]->GetCoeffs();
-
-            for (j = 0; j < bndCondExp[i]->GetNcoeffs(); ++j)
+            if (m_fields[nv]->GetBndConditions()[i]->GetBoundaryConditionType()
+                == SpatialDomains::eDirichlet)
             {
-                NekDouble sign =
-                    m_assemblyMap->GetBndCondCoeffsToGlobalCoeffsSign(
-                        bndcnt);
-                inout[bndMap[bndcnt++]] = sign * bndCoeffs[j];
+                const Array<OneD,const NekDouble> &bndCoeffs =
+                    bndCondExp[i]->GetCoeffs();
+
+                for (j = 0; j < bndCondExp[i]->GetNcoeffs(); ++j)
+                {
+                    NekDouble sign =
+                        m_assemblyMap->GetBndCondCoeffsToGlobalCoeffsSign(
+                            bndcnt);
+                    inout[bndMap[bndcnt++]] = sign * bndCoeffs[j];
+                }
+            }
+            else
+            {
+                bndcnt += bndCondExp[i]->GetNcoeffs();
             }
         }
     }